requirements.txt
Django==1.5.1
SQLAlchemy==0.8.2
argparse==1.2.1
django-admin-bootstrapped==0.4.1
django-crispy-forms==1.3.2
django-debug-toolbar==0.9.4
django-floppyforms==1.1
django-form-utils==0.3.1
django-jsonify==0.2.1
django-suit==0.2.2
django-widget-tweaks==1.3
djangorestframework==2.3.6
mimeparse==0.1.3
python-dateutil==2.1
pytz==2013b
six==1.3.0
wsgiref==0.1.2
